AttributeImportDraft to interface. #942
