+++ sphere-models/src/it/java/io/sphere/sdk/orders/commands/OrderImportCommandTest.java
@@ -5,7 +5,6 @@ import io.sphere.sdk.cartdiscounts.DiscountedLineItemPrice;
 import io.sphere.sdk.cartdiscounts.DiscountedLineItemPriceForQuantity;
 import io.sphere.sdk.carts.*;
 import io.sphere.sdk.channels.ChannelRole;
-import io.sphere.sdk.customers.CustomerFixtures;
 import io.sphere.sdk.models.Address;
 import io.sphere.sdk.models.DefaultCurrencyUnits;
 import io.sphere.sdk.models.LocalizedString;
@@ -15,7 +14,6 @@ import io.sphere.sdk.products.*;
 import io.sphere.sdk.products.attributes.Attribute;
 import io.sphere.sdk.products.attributes.AttributeAccess;
 import io.sphere.sdk.products.attributes.AttributeImportDraft;
-import io.sphere.sdk.products.search.ProductProjectionSearch;
 import io.sphere.sdk.shippingmethods.ShippingMethod;
 import io.sphere.sdk.shippingmethods.ShippingRate;
 import io.sphere.sdk.suppliers.TShirtProductTypeDraftSupplier;
@@ -32,7 +30,6 @@ import javax.money.MonetaryAmount;
 import java.math.BigDecimal;
 import java.time.ZonedDateTime;
 import java.util.List;
-import java.util.Locale;
 import java.util.Set;
 import java.util.UUID;
 import java.util.function.Consumer;
